requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Babel==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
seleniumbase==4.25.0
beautifulsoup4==4.12.3
lxml==5.1.0
googletrans==4.0.0rc1
deep-translator==1.11.4

